+++ src/main/java/org/kakaoshare/backend/domain/order/repository/query/OrderRepositoryCustom.java
@@ -3,7 +3,7 @@ package org.kakaoshare.backend.domain.order.repository.query;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.kakaoshare.backend.domain.rank.dto.RankResponse;
-import org.kakaoshare.backend.domain.rank.entity.TargetType;
+import org.kakaoshare.backend.domain.rank.util.TargetType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
