+++ src/main/java/org/kakaoshare/backend/domain/rank/controller/RankController.java
@@ -5,8 +5,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.kakaoshare.backend.domain.rank.dto.PriceRange;
 import org.kakaoshare.backend.domain.rank.dto.RankResponse;
-import org.kakaoshare.backend.domain.rank.entity.RankType;
-import org.kakaoshare.backend.domain.rank.entity.TargetType;
+import org.kakaoshare.backend.domain.rank.util.RankType;
+import org.kakaoshare.backend.domain.rank.util.TargetType;
 import org.kakaoshare.backend.domain.rank.service.RankService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
