+++ src/main/java/org/kakaoshare/backend/domain/rank/service/RankService.java
@@ -7,8 +7,8 @@ import lombok.RequiredArgsConstructor;
 import org.kakaoshare.backend.domain.order.repository.OrderRepository;
 import org.kakaoshare.backend.domain.rank.dto.PriceRange;
 import org.kakaoshare.backend.domain.rank.dto.RankResponse;
-import org.kakaoshare.backend.domain.rank.entity.RankType;
-import org.kakaoshare.backend.domain.rank.entity.TargetType;
+import org.kakaoshare.backend.domain.rank.util.RankType;
+import org.kakaoshare.backend.domain.rank.util.TargetType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
